sty: move confounds.py in the code folder and symlink it inside analysis
